Remove Tests for ReSwift Actions
